backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pymongo>=4.11.0
python-dotenv>=1.0.0
pydantic[email]>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
gunicorn>=23.0.0